pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skillsync"
version = "1.0.0"
description = "Sync your AI skills across Claude Code and OpenClaw"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "SkillSync Team", email = "team@skillsync.dev"}
]
keywords = ["ai", "claude", "openclaw", "skills", "sync", "github"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.0.0",
    "rich>=13.0.0",
    "watchdog>=3.0.0",
    "PyGithub>=2.0.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/skillsync"
Documentation = "https://skillsync.dev/docs"
Repository = "https://github.com/yourusername/skillsync"
Issues = "https://github.com/yourusername/skillsync/issues"

[project.scripts]
skillsync = "skillsync.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["skillsync*"]
